refactor(AppHeader): Split responsibilities
